+++ docs/CHANGELOG.md
@@ -19,6 +19,10 @@ Changes are grouped as follows:
 ### Fixed
 - The `neat.prepare.instances.relationships_as_edges()` no longer creates invalid identifiers for the edges.
 
+### Added
+- Fast serialization of NeatGraphStore
+- `neat.to.session(...)` and `neat.from.session(...)` methods to save and load a `NeatSession` object to and from a file.
+
 ### Improved
 - Better error message if `NeatSession(..., storage="oxigraph")` and the `oxigraph` package is not installed.